RequestKey

+++ common-ui-list/common-pr/src/main/java/com/storyteller_f/common_pr/Common.kt
@@ -23,5 +23,5 @@ val Float.dip: Float
 
 context(Context)
         @ExtFuncFlat(ExtFuncFlatType.V4, isContextReceiver = true)
-val Float.dipToInt: Float
-    get() = TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, this@dipToInt, resources.displayMetrics)
+val Float.dipToInt: Int
+    get() = TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, this@dipToInt, resources.displayMetrics).toInt()
